Delete run.bat
this is a specific method of running the file that automates a process that is already explained through the readme. users can create their own run.bat's, but i wouldnt include a specific way to run it in only a specific mode in the repository. its a localized utility, and not something that would translate well to the variety of uses this file has.